requirements.txt
python-telegram-bot==21.6
google-generativeai==0.8.3
anthropic==0.40.0
tweepy==4.14.0
requests==2.32.3
python-dotenv==1.0.1
Pillow
aiohttp
aiofiles==24.1.0
